backend/requirements.txt
fastapi
uvicorn
pydantic 
sqlalchemy
pymysql 
passlib[bcrypt]
python-jose 